chatbot-app/requirements.txt
# Core dependencies
flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
google-generativeai==0.3.2

# Additional dependencies for production
gunicorn==21.2.0
python-json-logger==2.0.7